refactored storage upload
